+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 2020-06-06 - 1.0.77
 - Adding `seperator` class for `vcl\ui\Node`
+- Updating `prototypes/ui/forms/util/Console` (better Alt+Click)
 
 ### 2020-06-02 - 1.0.76
 - App.js - Button ???